requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
pillow
qdrant-client
motor
aioredis
httpx
pytest
python-multipart
python-dotenv
